+++ motorized_fader_control/lib/EventBus.js
@@ -1,0 +1,17 @@
+// lib/EventBus.js
+class EventBus {
+    constructor() {
+      this.listeners = {};
+    }
+  
+    on(event, callback) {
+      if (!this.listeners[event]) this.listeners[event] = [];
+      this.listeners[event].push(callback);
+    }
+  
+    emit(event, data) {
+      (this.listeners[event] || []).forEach(cb => cb(data));
+    }
+  }
+
+  module.exports = EventBus;
