+++ motorized_fader_control/lib/services/AlbumService copy.js
@@ -1,0 +1,36 @@
+// services/AlbumService.js
+class AlbumService extends BaseService {
+    constructor(...args) {
+      super(...args);
+      
+      // Existing album cache integration
+      this.lastAlbumUri = null;
+      
+      this.eventBus.on('playback/update', (state) => {
+        if(this.shouldUpdateHardware(state)) {
+          const position = this.calculatePosition(state);
+          this.updateHardware(position);
+        }
+      });
+    }
+  
+    calculatePosition(state) {
+      // Preserve your existing progression logic
+      if(this.config.get('SEEK_TYPE') === 'album') {
+        return this.calculateAlbumProgression(state);
+      }
+      return (state.seek / state.duration) * 100;
+    }
+  
+    calculateAlbumProgression(state) {
+      // Integrate your existing album progression logic
+      if(this.stateCache.get('currentAlbum')?.uri !== state.uri) {
+        this.fetchAlbumInfo(state.uri);
+      }
+      
+      // ... rest of your existing album logic
+      //! realtime seek integration
+    }
+  }
+
+module.exports = AlbumService;
