+++ motorized_fader_control/lib/services/BaseService.js
@@ -1,0 +1,21 @@
+// services/BaseService.js
+class BaseService {
+    constructor(faderIdx, eventBus, stateCache) {
+      this.faderIdx = faderIdx;
+      this.eventBus = eventBus;
+      this.stateCache = stateCache;
+    }
+  
+    handleMove(position) {
+      /* To be overridden */
+    }
+  
+    updateHardware(position) {
+      this.eventBus.emit('hardware/update', {
+        fader: this.faderIdx,
+        position
+      });
+    }
+  }
+  
+module.exports = BaseService;
