+++ motorized_fader_control/lib/services/TrackService copy.js
@@ -1,0 +1,24 @@
+// services/TrackService.js
+class TrackService extends BaseService {
+  constructor(...args) {
+    super(...args);
+    
+    // Existing album cache integration
+    this.lastAlbumUri = null;
+    
+    this.eventBus.on('playback/update', (state) => {
+      if(this.shouldUpdateHardware(state)) {
+        const position = this.calculatePosition(state);
+        this.updateHardware(position);
+      }
+    });
+  }
+
+  calculatePosition(state) {
+
+  }
+
+  calculateTrackProgression(state) {
+
+  }
+}
